.ipynb_checkpoints/requirements-checkpoint.txt
datetime
tqdm
pandas
numpy
scikit-image
torchvision
scipy
scikit-learn
visdom
path
multitasking